% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiclusterExperiment.R
\docType{methods}
\name{addStrat,BiclusterExperiment,numeric-method}
\alias{addStrat,BiclusterExperiment,numeric-method}
\title{Add a BiclusterStrategy calculated with the given k and method}
\usage{
\S4method{addStrat}{BiclusterExperiment,numeric}(bce, k, method = c("als-nmf",
  "svd-pca", "snmf", "nipals-pca", "plaid", "spectral"), maxNa, duplicable,
  silent, ...)
}
\description{
Returns a BiclusterExperiment identical to \code{bce} with the addition of a
BiclusterStrategy accessible using \code{strategies()} or \code{getStrat()}.
}
\details{
The provided \code{method} is used to compute a number of biclusters,
sets comprising both samples and features. Matrix factorization methods will
store intermediate data in the \code{factors} slot of the BiclusterStrategy.
That intermediate data is then thresholded to yield the matrices in the
\code{clusteredSamples} and \code{clusteredFeatures} slots of the
BiclusterStrategy.
}
\section{Potential side effects}{

Due to requirements of various biclustering methods, this function may with
warning override user parameters. Also, if any elements of 
\code{abund(BiclusterExperiment)} are missing, the row and column
containing those elements may be removed with warning.
}

