% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addStrat.R
\docType{methods}
\name{addStrat}
\alias{addStrat}
\alias{addStrat,BiclusterExperiment,numeric-method}
\title{Add a BiclusterStrategy to a BiclusterExperiment}
\usage{
addStrat(bce, k, method = c("als-nmf", "svd-pca", "snmf", "nipals-pca",
  "plaid", "spectral"), duplicable = TRUE, silent = FALSE, ...)

\S4method{addStrat}{BiclusterExperiment,numeric}(bce, k,
  method = c("als-nmf", "svd-pca", "snmf", "nipals-pca", "plaid",
  "spectral"), duplicable = TRUE, silent = FALSE, ...)
}
\arguments{
\item{bce}{A \code{\link{BiclusterExperiment-class}} object to analyze}

\item{k}{The number of biclusters to find}

\item{method}{The biclustering algorithm}

\item{duplicable}{Makes biclustering deterministic}

\item{silent}{Suppresses warnings and messages}

\item{...}{Additional parameters to pass to the
\link[=bicluster-methods]{biclustering back-end}}
}
\value{
A copy of \code{bce} with a new BiclusterStrategy added to
 \code{BiclusterExperiment@strategies}
}
\description{
Returns a BiclusterExperiment identical to \code{bce} with an additional
BiclusterStrategy accessible using \code{strategies()} or \code{getStrat()}.
}
\details{
One of the \link[=bicluster-methods]{biclustering algorithms} is used to
compute a number of biclusters. Matrix factorization methods will store
matrix factors in the \code{factors} slot of the \code{BiclusterStrategy}.
The matrix factors will be thresholded to yield the binary matrices in the
\code{clusteredSamples} and \code{clusteredFeatures} slots of the
\code{BiclusterStrategy}.
}
\section{Methods (by class)}{
\itemize{
\item \code{bce = BiclusterExperiment,k = numeric}: Default method
}}

\section{Additional arguments}{
 \code{addStrat()} automatically manipulates
  certain back-end arguments as needed to return the desired number of
  biclusters (\code{k}). Back-end arguments besides those automatically
  manipulated can be provided by name to \code{addStrat()}. See
  \link{bicluster-methods} for a description of back-end arguments.
}

\section{Potential side effects}{
 Due to requirements of various biclustering
  methods, this function may override user parameters, with warning. Also, if
  any elements of \code{abund(BiclusterExperiment)} are missing, the row and
  column containing those elements may be removed, with warning.
}

\examples{
bce <- BiclusterExperiment(yeast_benchmark[[1]])
bce <- addStrat(bce, k = 2)

}
