% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bicluster.R
\name{als_nmf}
\alias{als_nmf}
\title{Alternating-least-squares non-negative matrix approximation}
\usage{
als_nmf(A, k, reps = 4L, maxIter = 100L, eps_conv = 1e-04,
  duplicable = TRUE, verbose = TRUE, ...)
}
\arguments{
\item{A}{the matrix to factorize}

\item{k}{the number of factors to calculate}

\item{reps}{the number of replications to choose from}

\item{maxIter}{the maximum number of least-squares steps}

\item{eps_conv}{convergence tolerance}

\item{duplicable}{fix the random seed internally}

\item{verbose}{Print the mean squared error every 10 iterations}
}
\value{
a \code{\link{genericFit-class}} object
}
\description{
Approximates a non-negative matrix as the product of two non-negative matrix
factors using the alternating-least-squares algorithm by Paatero and Tapper
(1994).
}
\details{
Factorization is performed \code{reps} times, then the result with the
minimum mean squared-error is returned. \code{als_nmf()} is fast for a small
number of biclusters, but running time rapidly increases with \code{k}.
}
