% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiclusterExperiment.R
\docType{methods}
\name{as.matrix,BiclusterExperiment-method}
\alias{as.matrix,BiclusterExperiment-method}
\title{Get abundance values in a BiclusterExperiment}
\usage{
\S4method{as.matrix}{BiclusterExperiment}(x)
}
\description{
Returns a numeric matrix
}
