% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcv.R
\name{auto_bcv}
\alias{auto_bcv}
\title{Perform bcv until convergence}
\usage{
auto_bcv(Y, ks, maxIter = 100, tol = (10^-4), bestOnly = TRUE,
  verbose = TRUE)
}
\arguments{
\item{Y}{the input matrix}

\item{maxIter}{maximum number of iterations}

\item{tol}{tolerance used to determine convergence}

\item{bestOnly}{if FALSE, both the predicted number of biclusters and a table
of result counts is returned}

\item{kLimit}{the maximum number of biclusters to consider}
}
\description{
Performs BCV until the the distribution of results has converged. Often this
requires less than 50 iterations.
}
\details{
The returned number of biclusters is the median of the results from
\code{maxIter} iterations.
}
\section{Deciding the maximum k}{

To tune kLimit, it might be helpful to run with \code{maxIter} around 10 and
\code{bestOnly = TRUE} to determine if there is an obvious upper bound on the
results.
}

