% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiclusterStrategy.R
\docType{methods}
\name{bcNames,BiclusterStrategy-method}
\alias{bcNames,BiclusterStrategy-method}
\title{Names of biclusters in this BiclusterStrategy}
\usage{
\S4method{bcNames}{BiclusterStrategy}(bcs, allBc = FALSE)
}
\description{
Returns a character vector of bicluster names, which can be used to reference
columns of \code{featureFactor(bcs)} and \code{clusteredFeatures(bcs)}, and
rows of \code{sampleFactor(bcs)} and code{clusteredSamples(bcs)}.
}
\details{
A character vector.
}
