% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{bcNames}
\alias{bcNames}
\title{Bicluster names}
\usage{
bcNames(bcs, allBc = FALSE)
}
\description{
Get the names of all biclusters in a BiclusterStrategy. This function is
mostly useful for graphical functions.
}
