% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcv.R
\name{bcv}
\alias{bcv}
\title{Perform bi-cross-validation}
\usage{
bcv(Y, ks, holdouts = 10L, duplicable = TRUE)
}
\arguments{
\item{Y}{the input matrix}

\item{ks}{the range of biclusters to evaluate}

\item{holdouts}{the number of holdouts to perform along each dimension of
matrix \code{Y}}
}
\description{
The number of biclusters yielding the lowest BCV value is the predicted best.
It is recommended to use auto_bcv to perform several replications of
bi-cross-validation, and use the median of the results as the predicted
number of biclusters.
}
\details{
A named vector of BCV values corresponding to the various numbers of
biclusters evaluated. The range of biclusters may be non-contiguous.
}
