% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bicluster.R
\name{bicluster}
\alias{bicluster}
\title{Biclustering algorithms}
\arguments{
\item{A}{a numeric matrix to bicluster}

\item{k}{the number of biclusters to report}

\item{duplicable}{causes biclutsering to be non-stochastic}

\item{verbose}{report details such as convergence and the underlying function
calls}
}
\description{
The \code{method} argument of \code{\link{addStrat}()} provides access to
functions from packages \code{\link{[NMF]}} and \code{\link{[biclust]}}.
Initially \code{addStrat()} uses the default parameters provided by the
respective developers, then progressively relaxes certain parameters if
needed to return the desired number of biclusters. Method-specific parameters
besides those automatically manipulated can be provided by name to
\code{addStrat()}. Please see respective documentation in other packages for
method-specific parameters.
}
\details{
\describe{ 
\item{als_nmf}{The alternating-least-squares non-negative matrix
factorization algorithm. The original algorithm by Paatero and Tapper (1994)
is run \code{reps} times, then the result that most accurately factorized the
matrix \code{A} is returned. \code{als_nmf} is fast for a small number of
bicluster, but running time rapidly increases with \code{k}.}
\item{svd_pca}{The singular vector decomposition algorithm. Each principal
component is interpreted as the degree of membership in a single bicluster.
The resulting score matrix is thresholded to binarize bicluster membership.
\code{svd_pca} is the fastest provided algorithm.}
\item{nipals_pca}{An iterative PCA algorithm that may tolerate missing data.
  Slower than \code{svd_pca}, but still faster than the other algorithms.}
\item{plaid}{The plaid algorithm}
\item{snmf}{Sparse non-negative matrix factorization. The core algorithm is
Paatero and Tapper's alternating least-squares (1994), and the
sparsity-inducing regularization factors were introduced by Kim and Park
(2007).}
\item{spectral}{The Spectral algorithm. For smaller matrices, the parameter
  \code{numberOfEigenvalues} is automatically set to enable finding the
  number of biclusters requested. The parameter \code{withinVar} is allowed
  to increase up to 10 times the smaller matrix dimension.}
}
}
\seealso{
\code{\link[NMF]{nmfAlgorithm.SNMF_R}}

\code{\link[biclust]{BCSpectral}}

\code{\link[biclust]{BCPlaid}}

\code{\link[nipals]{nipals}}
}
