% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiclusterExperiment.R
\docType{class}
\name{BiclusterExperiment-class}
\alias{BiclusterExperiment-class}
\alias{BiclusterExperiment,matrix-method}
\alias{BiclusterExperiment,ANY-method}
\alias{as.matrix,BiclusterExperiment-method}
\alias{names,BiclusterExperiment-method}
\alias{strategies,BiclusterExperiment-method}
\title{Class "BiclusterExperiment" for data and biclusters}
\usage{
\S4method{BiclusterExperiment}{matrix}(m, bcs = list(),
  phenoData = Biobase::annotatedDataFrameFrom(m, byrow = FALSE),
  featureData = annotatedDataFrameFrom(m, byrow = TRUE), pp = FALSE,
  maxNa = 0)

\S4method{BiclusterExperiment}{ANY}(m, bcs = list(),
  phenoData = Biobase::annotatedDataFrameFrom(m, byrow = FALSE),
  featureData = annotatedDataFrameFrom(m, byrow = TRUE), pp = FALSE,
  maxNa = 0)

\S4method{as.matrix}{BiclusterExperiment}(x)

\S4method{names}{BiclusterExperiment}(x)

\S4method{strategies}{BiclusterExperiment}(x)
}
\arguments{
\item{m}{The data on which derived BiclusterStrategy objects will be
calculated. Should be a matrix or a class coercible to matrix, with samples
as columns and features as rows}

\item{bcs}{May be a \code{list} of BiclusterStrategy objects that have been
computed from \code{m} previously}

\item{phenoData}{Metadata about the rows of \code{m}}

\item{featureData}{Metadata about the columns of \code{m}}

\item{pp}{Whether to preprocess \code{m} by removing rows and columns with
excessive missing data}

\item{maxNa}{If \code{pp}, rows and columns with over \code{maxNa} elements
missing are discarded}

\item{x}{A \code{BiclusterExperiment-class} object}
}
\description{
This class encapsulates data for one or more biclustering runs derived from
the same abundance data. Objects can be created using the
\code{\link{BiclusterExperiment}} constructor. A subclass of
\code{\link{eSet}}.
}
\section{Methods (by generic)}{
\itemize{
\item \code{BiclusterExperiment}: Default constructor method

\item \code{BiclusterExperiment}: Attempts to coerce \code{m} to a matrix before
encapsulating it in a \code{BiclusterExperiment}

\item \code{names}: Retrieves the names of any encapsulated
\code{\link{BiclusterStrategy-class}} objects.

\item \code{strategies}: Get/set a list of the BiclusterStrategy objects
contained by this BiclusterExperiment.
}}

\section{Slots}{

\describe{
\item{\code{assayData}}{Object of class \code{\link{AssayData-class}} in the
"list" storage mode. \code{AssayData} must have a matrix named
"abund" with rows representing features and columns representing samples.
Any other matrices in assayData are ignored by this package.}

\item{\code{strategies}}{A \code{\link{list}} of
\code{\link{BiclusterStrategy-class}} objects}

\item{\code{phenoData}}{See \code{\link{eSet-class}}.}

\item{\code{featureData}}{\code{\link{eSet-class}}.}

\item{\code{experimentData}}{\code{\link{eSet-class}}. Not accessed by any
functions in \code{\link{mfBiclust}}.}

\item{\code{annotation}}{\code{\link{eSet-class}}. Not accessed by any
functions in \code{\link{mfBiclust}}.}

\item{\code{protocolData}}{\code{\link{eSet-class}}. Not accessed by any
functions in \code{\link{mfBiclust}}.}
}}

\examples{
bce <- BiclusterExperiment(yeast_benchmark[[1]])

}
\seealso{
\code{\link{eSet-class}}

\code{\link{getStrat}}
}
