% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiclusterExperiment.R
\docType{methods}
\name{BiclusterExperiment,matrix-method}
\alias{BiclusterExperiment,matrix-method}
\title{Careful, for m, rows are samples and columns are features
eSet objects store assayData transposed: rows are features and columns are samples.
For this reason I wrote a getter that returns a matrix with rows as samples, columns as features.}
\usage{
\S4method{BiclusterExperiment}{matrix}(m, bcs, annot = data.frame(),
  bcv = FALSE)
}
\description{
Careful, for m, rows are samples and columns are features
eSet objects store assayData transposed: rows are features and columns are samples.
For this reason I wrote a getter that returns a matrix with rows as samples, columns as features.
}
