% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{biclusterGDS}
\alias{biclusterGDS}
\title{Perform biclustering on microarray data}
\usage{
biclusterGDS(bce, k = 5)
}
\description{
Returns a BiclusterExperiment object with Bicluster.X fields in phenoData.
Optionally, also a list containing bicluster assignments
}
