% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biclusterGUI.R
\docType{methods}
\name{biclusterGUI}
\alias{biclusterGUI}
\alias{biclusterGUI,BiclusterExperiment-method}
\title{Explore a BiclusterExperiment}
\usage{
biclusterGUI(obj)

\S4method{biclusterGUI}{BiclusterExperiment}(obj)
}
\description{
Opens a GUI to visualize analysis results contained in a BiclusterExperiment
object
}
\section{Methods (by class)}{
\itemize{
\item \code{BiclusterExperiment}: Open GUI for a BiclusterExperiment
}}

