% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{biclusterNumbers2membershipMatrices}
\alias{biclusterNumbers2membershipMatrices}
\title{Convert row and column numbers to a matrix representation}
\usage{
biclusterNumbers2membershipMatrices(biclusterNumber, m, k)
}
\arguments{
\item{biclusterNumber}{A list describing the rows and columns in each
bicluster}

\item{m}{The numeric matrix in which biclusters were computed}

\item{k}{Restricts the results to the first \code{k} biclusters}
}
\value{
a \code{\link{list}} of two matrices
}
\description{
This function converts the output from
\code{\link[biclust]{biclusternumber}()} into the membership matrix format.
The first returned membership matrix shows which rows are in which bicluster.
The second returned membership matrix shows which columns are in which
bicluster.
}
\examples{
require(biclust)
data <- matrix(rnorm(400), 20, 20)
data[12:16, 8:12] <- rnorm(25, 3, 0.3)
set.seed(1)
biclusters <- biclust(data, BCPlaid(), back.fit = 2, shuffle = 3,
 fit.model = ~m + a + b, iter.startup = 5, iter.layer = 30, verbose = TRUE)
biclusterNumbers <- biclusternumber(biclusters)
biclusterNumbers2membershipMatrices(biclusterNumbers, data, 1)
}
