% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiclusterStrategy.R
\docType{class}
\name{BiclusterStrategy-class}
\alias{BiclusterStrategy-class}
\alias{nclust,BiclusterStrategy-method}
\title{Class "BiclusterStrategy" for biclusters}
\usage{
\S4method{nclust}{BiclusterStrategy}(bcs)
}
\description{
This class encapsulates bicluster results for one biclustering algorithm, one
thresholding algorithm, and one quantity of biclusters.
}
\section{Methods (by generic)}{
\itemize{
\item \code{nclust}: Number of clusters in a BiclusterStrategy
}}

