% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cancer_benchmark}
\alias{cancer_benchmark}
\title{Gene expression datasets for benchmarking classification performance}
\format{A list of 13 datasets where each element contains:
\describe{
  \item{data}{a dataframe of expression values, unmodified from biclustlib}
  \item{labels}{a binary matrix where rows are pre-classified samples and
    columns are named cancer categories}
}}
\source{
\url{https://github.com/padilha/biclustlib/tree/master/}
}
\usage{
cancer_benchmark
}
\description{
Microarray and RNAseq datasets from biclustlib, originally published by
de Souto MC, Costa IG, de Araujo DS, Ludermir TB, Schliep A in 2008.
}
\keyword{datasets}
