% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/containerGenerics.R
\name{clean}
\alias{clean}
\title{Clean a matrix}
\usage{
clean(object, cleanParam = 0, dimsRemain = FALSE)
}
\description{
Returns an object of the same type as \code{object}, containing all columns 
in \code{object} for which the fraction of missing values is less than
\code{cleanParam}
}
