% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/BiclusterStrategy.R
\docType{methods}
\name{clusteredFeatures}
\alias{clusteredFeatures}
\alias{clusteredFeatures,BiclusterStrategy-method}
\title{Bicluster-feature matrix}
\usage{
clusteredFeatures(bcs, allBc = FALSE)

\S4method{clusteredFeatures}{BiclusterStrategy}(bcs, allBc = FALSE)
}
\arguments{
\item{bcs}{A BiclusterStrategy}

\item{allBc}{TRUE to return data from all stored biclusters}
}
\value{
A logical matrix
}
\description{
Returns a logical matrix \eqn{A_{m x k}} where element \eqn{A_{i,j}} is TRUE
if feature \eqn{i} is in bicluster \eqn{j}. Elsewhere in \code{mfBiclust}
documentation, this format is called a "membership matrix".
}
\section{Methods (by class)}{
\itemize{
\item \code{BiclusterStrategy}: method for BiclusterStrategy objects
}}

\examples{
bce <- BiclusterExperiment(yeast_benchmark[[1]])
bce <- addStrat(bce, k = 2, method = "als-nmf")
bcs <- getStrat(bce, 1)
clusteredFeatures(bcs)

}
