% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiclusterStrategy.R
\name{clusteredSamples}
\alias{clusteredSamples}
\title{Sample-bicluster clustering matrix}
\usage{
clusteredSamples(bcs)
}
\description{
A binary matrix showing which samples are members of which biclusters.
}
