% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{createAnnots}
\alias{createAnnots}
\title{Create annotations dataframe for heatmaps}
\usage{
createAnnots(x, names, strategy = "", phenoLabels = c(),
  biclustLabels = c())
}
\arguments{
\item{x}{a BiclusterExperiment}

\item{names}{the names of entities annotated- required}

\item{strategy}{a strategy in x. Required whenever length(biclustLabels) > 0}

\item{phenoLabels}{any phenotype labels in x}

\item{biclustLabels}{any bicluster labels in x}
}
\description{
Checks phenoLabels and biclustLabels for validity and then joins the 
corresponding data extracted from the provided BiclusterExperiment.
}
\details{
Annotation tracks are converted to named dataframe columns.
}
