% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{duplicable}
\alias{duplicable}
\title{Use a string to set the random seed}
\usage{
duplicable(str)
}
\description{
Returns the output of .Random.seed. Call
\code{on.exit(assign(".Random.seed", oldSeed, envir=globalenv()), add = TRUE)
} to restore the random seed.
}
