% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{error.bar}
\alias{error.bar}
\title{Error bars}
\usage{
error.bar(x, y, upper, lower = upper, length = 0.1, ...)
}
\arguments{
\item{x}{the result of a call to barplot()}

\item{y}{the ys provided to barplot()}

\item{upper}{vector of lengths of upper error bars}

\item{lower}{vector of lengths of lower error bars; by default, same as 
upper}
}
\description{
Error bars
}
