% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphical.R
\docType{methods}
\name{factorHeatmap,BiclusterExperiment,BiclusterStrategy-method}
\alias{factorHeatmap,BiclusterExperiment,BiclusterStrategy-method}
\title{Factor matrix heatmap}
\usage{
\S4method{factorHeatmap}{BiclusterExperiment,BiclusterStrategy}(bce, bcs,
  type = c("score", "loading"), phenoLabels = c(), biclustLabels = c(),
  ordering = c("input", "distance", "cluster"), colNames = FALSE)
}
\description{
Display scores for all clusters in one heatmap
}
