% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphical.R
\name{factorHeatmap}
\alias{factorHeatmap}
\title{Plot a heatmap showing bicluster membership of samples or features}
\usage{
factorHeatmap(bce, bcs, type, ordering = "input", ...)
}
\arguments{
\item{bce}{A BiclusterExperiment object}

\item{bcs}{The name or index of a BiclusterStrategy contained by \code{bce},
or the BiclusterStrategy object itself}

\item{type}{either "score" for feature-bicluster membership or "loading" for
sample-bicluster membership}

\item{ordering}{\code{ordering = "distance"} to reorder by Euclidean distance
on \code{as.matrix(BiclusterExperiment)}. \code{ordering = "cluster"} to
reorder by Euclidean distance on the appropriate bicluster-membership
matrix in \code{bcs}.}

\item{phenoLabels}{an optional character vector of labels to annotate. If \code{type = "score"},
\code{phenoLabels} should be column names of \code{Biobase::phenoData(bce)}}

\item{biclustLabels}{an optional character vector of labels to annotate.
Should be elements of \code{bcNames(bcs)}. Both \code{phenoLabels} and
\code{biclustLabels} may be specified.}

\item{colnames}{if \code{TRUE}, labels the samples/features}
}
\description{
Uses the data from the \code{factors} slot of a BiclusterStrategy to show
bicluster membership across all samples or features.
}
