% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{fuzzyFeatures}
\alias{fuzzyFeatures}
\title{Score matrix}
\usage{
fuzzyFeatures(bcs, allBc = FALSE)
}
\description{
For a data matrix M x N factorized to produce k biclusters, the score matrix is M x k.
'
}
