% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiclusterStrategy.R
\name{generateThresholdMatrix}
\alias{generateThresholdMatrix}
\title{Combine threshold values and names into a matrix}
\usage{
generateThresholdMatrix(thresholds, matrix, biclustNames)
}
\arguments{
\item{thresholds}{either a numeric matrix, a numeric vector, or a character
vector}

\item{matrix}{the target matrix, whose columns will be thresholded}

\item{biclustNames}{names of the threshold matrix rows}
}
\description{
Thresholds may contain a vector of the names of desired threshold algorithms.
Alternatively, a numeric-mode object may be supplied. A single numeric
will be applied to all clusters. A vector of numerics, if the same size as
k, will be assumed to have a 1:1 relation with k. A matrix of numerics,
if k x Y, will be assumed to be a matrix where each row k contains Y
thresholds to plot for bicluster k. This argument's dimensions will be
checked for compatibility with the target matrix.
}
\details{
A matrix where each column is a series of thresholds. Each column is
named by the threshold algorithm, if specified. Each row is named by the
provided biclustNames.
}
