% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperClasses.R
\docType{class}
\name{genericFactorization-class}
\alias{genericFactorization-class}
\title{A pair of matrix factors}
\description{
Two matrices that can be thresholded to determine which samples and
features are members of which biclusters. After running a
matrix-factorization biclustering method, such as SVD-PCA, ALS-NMF, or SNMF,
these matrices will be score and loading matrices.
}
\section{Slots}{

\describe{
\item{\code{W}}{an m x k matrix}

\item{\code{H}}{a k x n matrix}
}}

