% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperClasses.R
\docType{class}
\name{genericFit-class}
\alias{genericFit-class}
\title{A biclustering algorithm result}
\description{
Contains the results and metadata from running a biclustering algorithm.
}
\section{Slots}{

\describe{
\item{\code{fit}}{a \code{\link{genericFactorization}} object}

\item{\code{method}}{the name of the biclustering algorithm used}
}}

