% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiclusterExperiment.R
\name{getStrat}
\alias{getStrat}
\title{Access a BiclusterStrategy contained by a BiclusterExperiment}
\usage{
getStrat(bce, id)
}
\arguments{
\item{bce}{A BiclusterExperiment to access}

\item{id}{Either the integer index or the name of the BiclusterStrategy to 
get}
}
\description{
Returns one BiclusterStrategy. Alternatively, a list of BiclusterStrategy
objects contained by a BiclusterExperiment can be retrieved using
\code{strategies(bce)}.
}
