% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiclusterExperiment.R
\docType{methods}
\name{heatmapFactor,BiclusterExperiment-method}
\alias{heatmapFactor,BiclusterExperiment-method}
\title{Factor matrix heatmap}
\usage{
\S4method{heatmapFactor}{BiclusterExperiment}(obj, strategy = "",
  type = c("score", "loading"), phenoLabels = c(), biclustLabels = c(),
  ordering = c("input", "distance", "cluster"), colNames = FALSE)
}
\description{
Display scores for all clusters in one heatmap
}
