% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiclusterExperiment.R
\name{heatmapFactor}
\alias{heatmapFactor}
\title{#### Cluster stability plot ####################################################
#' @export
setGeneric("plotClustStab", signature = "obj", function(obj, ...) {
  standardGeneric("plotClustStab")
})}
\usage{
heatmapFactor(obj, ...)
}
\description{
#' Cluster stability plot
#'
#' Plots the stability of clusters in the given BiclusterExperiment.
#'
#' @rdname plotClustStab
#' @aliases plotClustStab
setMethod(
  "plotClustStab", signature(obj = "BiclusterExperiment"),
  function(obj, clusters) {
    if (length(obj@strategies) == 0) {
      warning(paste0("Please calculate clusters first!"))
      return(obj)
    }
    
    # calculate consensus too?
    
    # calculate stability of the clusters check if there are more than 1 k value in ks range
    stability <- NULL
    # stability <- calculate_stability(metadata(object)$sc3$consensus, k)
    stability <- abs(runif(n = clusters, min = 0, max = 1))
    
    d <- data.frame(Cluster = factor(1:length(stability)), Stability = stability)
    ggplot2::ggplot(d, ggplot2::aes_string(x = "Cluster", y = "Stability")) +
      ggplot2::geom_bar(stat = "identity") + ggplot2::ylim(0, 1) +
      ggplot2::labs(x = "Cluster", y = "Stability Index") + ggplot2::theme_bw()
  }
)
}
