% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphical.R
\name{matrixHeatmap}
\alias{matrixHeatmap}
\title{Quickly preview a matrix as a heatmap}
\usage{
matrixHeatmap(m)
}
\arguments{
\item{m}{A numeric matrix}
}
\value{
Invisibly returns NULL
}
\description{
Uses R's \code{image()} function to visualize a matrix.
}
\examples{
matrixHeatmap(yeast_benchmark[[1]])

}
