% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfbc.R
\name{mfbc}
\alias{mfbc}
\title{Analyze a matrix and display results graphically}
\usage{
mfbc(m, ...)
}
\arguments{
\item{x}{The input matrix}

\item{annotations}{A dataframe, where each column contains a set of labels on
the rows of \code{x}}

\item{ks}{the number of clusters to create, can be atomic or a vector}

\item{bicluster}{the biclustering algorithm}

\item{scoreThresh}{the thresholding algorithm to use for scores. Ignored if
\code{bicluster} is \code{"plaid"} or \code{"bimax"}. Currently "snmf/l"
and "pca" are implemented}

\item{loadingThresh}{the thresholding algorithm to use for loading. See note
on \code{scoreThresh}}

\item{bcv}{Perform bi-cross-validation? Not yet implemented.}
}
\description{
Runs all steps of the biclustering pipeline.
}
\details{
\code{mfbc} is the fastest way to analyze data. The only required option is
the input matrix; the default pipeline uses non-negative matrix factorization
with otsu thresholding. To compare multiple pipelines, manual construction of
a \code{\link{BiclusterExperiment}} object is recommended.
}
\seealso{
\code{\link{BiclusterExperiment}}
}
