% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfBiclust.R
\docType{package}
\name{mfBiclust}
\alias{mfBiclust}
\alias{mfBiclust-package}
\title{Matrix-factorization biclustering toolbox}
\description{
Biclustering using PCA and non-negative matrix factorization
}
\details{
This package contains functions and classes for creating and analyzing
biclustering pipelines. Currently ALS, SVD, NIPALS, SNMF, Spectral, and
Plaid are supported.
}
\section{Functions}{

  \code{\link{biclusterGUI}}
  \code{\link{BiclusterExperiment-class}}
  \code{\link{addStrat}}
  \code{\link{BiclusterStrategy-class}}
}

\references{
\href{https://doi.org/10.1016/j.ymeth.2018.02.004}{Bi-clustering
  of metabolic data using matrix factorization tools}
}
\author{
Quan Gu, <quan.gu@glasgow.ac.uk>
  Maintainer: Jonathan Lim, <jlim7@comcast.net>
}
