% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiclusterStrategy.R
\docType{methods}
\name{name,BiclusterStrategy-method}
\alias{name,BiclusterStrategy-method}
\title{Name of a BiclusterStrategy}
\usage{
\S4method{name}{BiclusterStrategy}(bcs)
}
\description{
Get this BiclusterStrategy's display-friendly name. If it does not have a name, computes a
string containing the biclustering algorithm, thresholding algorithms, and
number of biclusters in the given BiclusterStrategy.
}
\details{
This function may not be used to modify a BiclusterStrategy's name.
}
