% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{name}
\alias{name}
\title{Name a BiclusterStrategy}
\usage{
name(bcs)
}
\arguments{
\item{bcs}{a \code{\linkS4class{BiclusterStrategy}} object}
}
\description{
Gets a display-friendly name of a BiclusterStrategy that includes its
biclustering algorithm, its thresholding algorithm, and the number of
biclusters.
}
