% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiclusterExperiment.R
\docType{methods}
\name{names,BiclusterExperiment-method}
\alias{names,BiclusterExperiment-method}
\title{Names of BiclusterStrategies in this BiclusterExperiment}
\usage{
\S4method{names}{BiclusterExperiment}(x)
}
\description{
Names of BiclusterStrategies in this BiclusterExperiment
}
