% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiclusterStrategy.R
\docType{methods}
\name{names,BiclusterStrategy-method}
\alias{names,BiclusterStrategy-method}
\title{Names of biclusters in this BiclusterStrategy}
\usage{
\S4method{names}{BiclusterStrategy}(x)
}
\description{
Names of biclusters in this BiclusterStrategy
}
