% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiclusterStrategy.R
\docType{methods}
\name{nclust,BiclusterStrategy-method}
\alias{nclust,BiclusterStrategy-method}
\title{Cluster count accessor}
\usage{
\S4method{nclust}{BiclusterStrategy}(bcs)
}
\description{
Get the number of biclusters created in this BiclusterStrategy.
}
\details{
This function may not be used to modify a BiclusterStrategy's number of biclusters.
}
