% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/containerGenerics.R
\docType{methods}
\name{nclust,list-method}
\alias{nclust,list-method}
\title{nclust convenience method}
\usage{
\S4method{nclust}{list}(bcs)
}
\description{
Helper in case nclust is called on a list containing one BiclusterStrategy
}
