% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{nclust}
\alias{nclust}
\title{Number of biclusters}
\usage{
nclust(bcs, allBc = FALSE)
}
\arguments{
\item{bcs}{A BiclusterStrategy}

\item{allBc}{TRUE to return the total number of biclusters that were found
during the original call to \code{\link{addStrat}}}
}
\value{
integer
}
\description{
Returns the number of biclusters in a \code{\link{BiclusterStrategy-class}}
object
}
\examples{
bce <- BiclusterExperiment(yeast_benchmark[[1]])
bce <- addStrat(bce, k = 2)
bcs <- getStrat(bce, 1)
nclust(bcs)

}
