% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiclusterStrategy.R
\name{otsuHelper}
\alias{otsuHelper}
\title{Calculate Otsu thresholds on each column of a matrix}
\usage{
otsuHelper(matrix)
}
\arguments{
\item{matrix}{the target matrix, whose columns will be thresholded}
}
\value{
numeric vector
}
\description{
Calculate Otsu thresholds on each column of a matrix
}
\examples{
m <- matrix(rnorm(400), 40, 10)
otsuHelper(m)
}
