% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{overlap}
\alias{overlap}
\title{Calculate overlaps between every pair of biclusters.}
\usage{
overlap(BiclusterRows, BiclusterCols, matrix = FALSE)
}
\arguments{
\item{BiclusterRows}{a list, each element containing row indices in the
bicluster}

\item{BiclusterCols}{a list, each element containing column indices in the
bicluster}

\item{matrix}{if true, returns a matrix where element i, j is the proportion
of bicluster i overlapping with bicluster j}
}
\description{
Calculate overlaps between every pair of biclusters.
}
