% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{pcaWrapper}
\alias{pcaWrapper}
\title{Wrapper for prcomp}
\usage{
pcaWrapper(m, k)
}
\arguments{
\item{m}{the target matrix}

\item{k}{the number of principal components}
}
\description{
Returns a \code{\link{genericFit}} object.
}
