% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiclusterExperiment.R
\docType{methods}
\name{plot,BiclusterExperiment,ANY-method}
\alias{plot,BiclusterExperiment,ANY-method}
\title{Abundance heatmap}
\usage{
\S4method{plot}{BiclusterExperiment,ANY}(x, logBase = 0, phenoLabels = c(),
  biclustLabels = c(), ordering = c("input", "distance", "cluster"),
  strategy = "", rowNames = FALSE, colNames = FALSE)
}
\description{
Plot a heatmap from values in a BiclusterExperiment object.
}
\details{
This method implements /code{pheatmap::pheatmap} for BiclusterExperiment objects.

Strategy and cluster must be given if ordering = "cluster" or annotateCluster = TRUE.
}
