% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{plot}
\alias{plot}
\title{Expression plot}
\usage{
plot(x, y, ...)
}
\arguments{
\item{x}{A \code{link{BiclusterExperiment-class}} object}

\item{y}{Not used in the \code{BiclusterStrategy} method}

\item{...}{Additional arguments:\describe{
  \item{logBase}{If TRUE, every assay value is transformed by the natural log}
  \item{phenoLabels}{Display phenotype labels along the samples?}
  \item{biclustLabels}{Display bicluster labels along the samples and features?}
  \item{ordering}{If "input", samples and features are not reordered. If
  "distance", sampels and features are reordered based on Euclidean distance.
  If "cluster", samples and features are reordered based on bicluster
  membership.}
  \item{strategy}{The BiclusterStrategy to read biclusters from, if bicluster-
 based information is to be plotted.}
  \item{rowNames}{TRUE to show feature names}
  \item{colNames}{TRUE to show sample names}
}}
}
\value{
a \code{\link[pheatmap]{pheatmap}-class} object
}
\description{
Create a heatmap from assay data in a BiclusterExperiment object. A BiclusterStrategy and bicluster must be named if ordering = "cluster" or
annotateCluster = TRUE.
}
\examples{
bce <- BiclusterExperiment(cancer_benchmark[[1]]$data,
                           phenoData = cancer_benchmark[[1]]$labels)
phenoData(bce)
plot(bce, phenoLabels = "phenotype")
}
