% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiclusterExperiment.R
\docType{methods}
\name{plotClustStab,BiclusterExperiment-method}
\alias{plotClustStab,BiclusterExperiment-method}
\alias{plotClustStab}
\title{Cluster stability plot}
\usage{
\S4method{plotClustStab}{BiclusterExperiment}(obj, clusters)
}
\description{
Plots the stability of clusters in the given BiclusterExperiment.
}
