% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiclusterExperiment.R
\docType{methods}
\name{plotDist,BiclusterExperiment-method}
\alias{plotDist,BiclusterExperiment-method}
\alias{plotDist}
\title{Distance heatmap}
\usage{
\S4method{plotDist}{BiclusterExperiment}(x, distType = c("euclidean",
  "pearson"), phenoLabels = c(), biclustLabels = c(),
  ordering = c("input", "distance", "cluster"), strategy = "",
  rowColNames = FALSE)
}
\description{
Plot a heatmap from distance values in the BiclusterExperiment object
}
