% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiclusterExperiment.R
\docType{methods}
\name{plotMarkers,BiclusterExperiment-method}
\alias{plotMarkers,BiclusterExperiment-method}
\title{cluster must be numeric}
\usage{
\S4method{plotMarkers}{BiclusterExperiment}(obj, threshold, strategy = "",
  bicluster = "Bicluster.1", ordering = c("input", "cluster"))
}
\description{
cluster must be numeric
}
