% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiclusterExperiment.R
\docType{methods}
\name{plotSamples,BiclusterExperiment-method}
\alias{plotSamples,BiclusterExperiment-method}
\alias{plotSamples}
\title{Score plot}
\usage{
\S4method{plotSamples}{BiclusterExperiment}(obj, thresholds, strategy = "",
  bicluster = "Bicluster.1", ordering = c("input", "distance", "cluster"))
}
\description{
Plot cluster membership scores
}
