% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphical.R
\docType{methods}
\name{plotThreshold}
\alias{plotThreshold}
\alias{plotThreshold,BiclusterExperiment,BiclusterStrategy-method}
\title{Threshold plot}
\usage{
plotThreshold(bce, bcs, type, bicluster, ...)

\S4method{plotThreshold}{BiclusterExperiment,BiclusterStrategy}(bce, bcs,
  type = c("score", "loading"), bicluster = "Bicluster.1",
  thresholds = NULL, ordering = c("input", "distance", "cluster"),
  xlabs = FALSE)
}
\arguments{
\item{thresholds}{A single numeric, vector, or matrix of thresholds. See 
documentation}
}
\description{
Plot cluster membership values and thresholding
}
