% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/containerGenerics.R
\name{pred}
\alias{pred}
\title{Bicluster prediction accessor}
\usage{
pred(bcs)
}
\description{
Get a binary matrix coding for bicluster membership. Rows represent samples and columns represent biclusters.
}
