% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simdata}
\alias{simdata}
\title{Simulated bicluster datasets}
\format{A list of 4 matrices named:
\describe{
 \item{nonoverlap3}{}
 \item{overlap3}{}
 \item{nonoverlap10}{}
 \item{overlap10}{}
 }}
\usage{
simdata
}
\description{
Simulated matrices containing biclusters and mild normally-distributed noise.
Generated using \code{genSimData}.
}
\keyword{datasets}
