% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{sizes}
\alias{sizes}
\title{Calculate the size of all biclusters}
\usage{
sizes(biclusterRows, biclusterCols)
}
\description{
Provide biclusters as a list of rows in each bicluster, and a list of columns
in each bicluster.
}
