% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bicluster.R
\name{snmf}
\alias{snmf}
\title{Wrapper for NMF::nmf(method = "snmf/l")}
\usage{
snmf(m, k, beta = 0.01, verbose = FALSE, duplicable = FALSE)
}
\arguments{
\item{m}{the target matrix}

\item{k}{the size of the reduced dimension}

\item{beta}{the starting beta}
}
\description{
This wrapper explicitly enables use of the non-regularized alternating least
squares algorithm devised by Paatero and Tapper (1994), with optional
parameters providing access to coefficients of the regularization factors
introduced by Kim and Park (2007)
}
\details{
An \code{NMFfit} object
}
