% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{snmfWrapper}
\alias{snmfWrapper}
\title{Wrapper for NMF::nmf(method = "snmf/l")}
\usage{
snmfWrapper(m, k, beta = 0.01)
}
\arguments{
\item{m}{the target matrix}

\item{k}{the size of the reduced dimension}

\item{beta}{the starting beta}
}
\description{
This wrapper lowers the beta (sparsity) parameter if the ALS algorithm gets
/"stuck/" with an empty row or column too many times.
}
\details{
An \code{NMFfit} object
}
