% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bicluster.R
\name{svd_pca}
\alias{svd_pca}
\title{Wrapper for prcomp}
\usage{
svd_pca(A, k, duplicable = NULL, ...)
}
\arguments{
\item{k}{the number of principal components}

\item{m}{the target matrix}
}
\description{
Returns a \code{\link{genericFit}} object.
}
