% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{threshAlgo}
\alias{threshAlgo}
\title{Algorithm used to calculate a threshold}
\usage{
threshAlgo(bcs)
}
\arguments{
\item{bcs}{A \code{BiclusterStrategy-class} object}
}
\value{
A character string
}
\description{
Returns \code{\link{BiclusterStrategy-class}@threshAlgo}
}
\examples{
bce <- BiclusterExperiment(yeast_benchmark[[1]])
bce <- addStrat(bce, k = 2, method = "als-nmf")
bcs <- getStrat(bce, 1)
threshAlgo(bcs)
}
