% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\docType{methods}
\name{threshold,matrix-method}
\alias{threshold,matrix-method}
\title{Apply threshold to a score or loading matrix}
\usage{
\S4method{threshold}{matrix}(m, th)
}
\description{
Returns a binary matrix of the same size as \code{m} where all elements over
the threshold are 1.
}
