% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/BiclusterStrategy.R
\docType{methods}
\name{threshold}
\alias{threshold}
\alias{threshold,matrix,numeric-method}
\title{Apply thresholds to a score or loading matrix}
\usage{
threshold(m, th, MARGIN = 2)

\S4method{threshold}{matrix,numeric}(m, th, MARGIN = 2)
}
\arguments{
\item{m}{A numeric matrix}

\item{th}{A numeric vector with the same number of elements as
\code{dim(m)[[MARGIN]]}}

\item{MARGIN}{1 to threshold rows; 2 to threshold columns.}
}
\value{
A binary numeric matrix of the same dimensions as \code{m}
}
\description{
Returns a binary matrix of the same size as \code{m} where all elements over
the threshold are 1.
}
\section{Methods (by class)}{
\itemize{
\item \code{m = matrix,th = numeric}: Apply threshold to a score or loading matrix
}}

\examples{
m <- matrix(rnorm(100), 10, 10)
m
threshold(m = m, th = rep(0, 10), MARGIN = 1)

}
