% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{union.biclust}
\alias{union.biclust}
\title{Create a binary matrix containing a union of biclusters}
\usage{
union.biclust(RowxBiclust, BiclustxCol, choose = rep(TRUE, ncol(RowxBiclust)))
}
\arguments{
\item{RowxBiclust}{a boolean score matrix}

\item{BiclustxCol}{a boolean loading matrix}

\item{choose}{a vector of indexes of biclusters, should be the same length as
the width of RowxBiclust}
}
\description{
Create a binary matrix containing a union of biclusters
}
