% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/BiclusterExperiment.R
\docType{methods}
\name{wipe}
\alias{wipe}
\alias{wipe,BiclusterExperiment-method}
\title{Wipe biclustering results}
\usage{
wipe(bce)

\S4method{wipe}{BiclusterExperiment}(bce)
}
\arguments{
\item{bce}{A \code{BiclusterExperiment-class} object}
}
\value{
A \code{BiclusterExperiment-class} object
}
\description{
Returns a \code{\link{BiclusterExperiment-class}} object with all
BiclusterStrategies removed.
}
\examples{
bce <- BiclusterExperiment(yeast_benchmark[[1]])
bce <- addStrat(bce, k = 2, method = "als-nmf")
wipe(bce)
}
