% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiclusterExperiment.R
\name{wipeExcept}
\alias{wipeExcept}
\title{Removes all BiclusterStrategy objects except one}
\usage{
wipeExcept(bce, bcs)
}
\description{
Returns a BiclusterExperiment encapsulating only the named BiclusterStrategy.
Can be useful to store the final results.
}
