% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiclusterExperiment.R
\name{addStrat}
\alias{addStrat}
\title{Add a BiclusterStrategy to a BiclusterExperiment}
\usage{
addStrat(bce, k, method = c("als-nmf", "svd-pca", "snmf", "nipals-pca"),
  cleanParam = 0)
}
\description{
Due to requirements of various biclustering methods, this function may warn
that the chosen biclustering method was overridden.
}
\details{
Also, if any matrix elements of abund(BiclusterExperiment) are missing, then
the BiclusterExperiment returned is not guaranteed to have the same number of
samples and features.
}
