% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiclusterExperiment.R
\name{addStrat}
\alias{addStrat}
\title{Add a BiclusterStrategy to a BiclusterExperiment}
\usage{
addStrat(bce, k, method = c("als-nmf", "svd-pca", "snmf", "nipals-pca",
  "plaid", "spectral"), duplicable = TRUE, silent = FALSE, ...)
}
\description{
Returns a BiclusterExperiment identical to \code{bce} with the addition of a
BiclusterStrategy accessible using \code{strategies()} or \code{getStrat()}.
}
