% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addStrat.R
\name{addStrat}
\alias{addStrat}
\title{Add a BiclusterStrategy to a BiclusterExperiment}
\usage{
addStrat(bce, k, method = c("als-nmf", "svd-pca", "snmf", "nipals-pca",
  "plaid", "spectral"), duplicable = TRUE, silent = FALSE, ...)
}
\description{
Returns a BiclusterExperiment identical to \code{bce} with an additional
BiclusterStrategy accessible using \code{strategies()} or \code{getStrat()}.
}
\details{
Argument \code{method} is used to compute a number of biclusters. Matrix
factorization methods will store matrix factors in the \code{factors} slot of
the \code{BiclusterStrategy}. The matrix factors will be thresholded to yield the
binary matrices in the \code{clusteredSamples} and \code{clusteredFeatures}
slots of the \code{BiclusterStrategy}.
}
\section{Potential side effects}{
 Due to requirements of various biclustering
  methods, this function may override user parameters, with warning. Also, if
  any elements of \code{abund(BiclusterExperiment)} are missing, the row and
  column containing those elements may be removed, with warning.
}

\examples{
bce <- BiclusterExperiment(yeast_benchmark[[1]])
addStrat(bce, k = 3)
biclusterGUI(bce)
}
