% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiclusterStrategy.R
\docType{methods}
\name{bcNames,BiclusterStrategy-method}
\alias{bcNames,BiclusterStrategy-method}
\title{Names of biclusters in this BiclusterStrategy}
\usage{
\S4method{bcNames}{BiclusterStrategy}(bcs, allBc = FALSE)
}
\description{
Names of biclusters in this BiclusterStrategy
}
