% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{bcNames}
\alias{bcNames}
\title{Names of biclusters in this BiclusterStrategy}
\usage{
bcNames(bcs, allBc = FALSE)
}
\arguments{
\item{bcs}{A BiclusterStrategy}

\item{allBc}{TRUE to return data from all stored biclusters}
}
\value{
A character vector.
}
\description{
Returns a character vector of bicluster names, which can be used to reference
columns of \code{featureFactor(bcs)} and \code{clusteredFeatures(bcs)}, and
rows of \code{sampleFactor(bcs)} and code{clusteredSamples(bcs)}. Mostly
useful for graphical functions.
}
\examples{
bce <- BiclusterExperiment(yeast_benchmark[[1]])
bce <- addStrat(bce, k = 2, method = "als-nmf")
bcs <- getStrat(bce, 1)
bcNames(bcs)
}
