% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcv.R
\name{bcv}
\alias{bcv}
\title{Calculate bi-cross-validation}
\usage{
bcv(Y, ks = seq_len(min(nrow(Y), ncol(Y)) - 1), holdouts = 10L,
  duplicable = FALSE, interactive = TRUE)
}
\arguments{
\item{Y}{the input matrix}

\item{ks}{a vector of bicluster quantities to evaluate}

\item{holdouts}{the number of row and column partitions. The true number of
holdouts will be \code{holdouts} ^ 2.}

\item{duplicable}{fix the random seed internally}

\item{interactive}{prompt before running if \code{Y} is missing values}
}
\value{
A named vector of BCV values corresponding to the various numbers of
  biclusters evaluated.
}
\description{
Calculates the bi-cross-validation of a Singular value decmoposition of
matrix \code{Y}.
}
\details{
It is recommended to use auto_bcv to perform and analyze several replications
of bi-cross-validation.
}
\examples{
bcv(yeast_benchmark[[1]])

}
\seealso{
\code{\link{auto_bcv}()}
}
