% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiclusterExperiment.R
\docType{class}
\name{BiclusterExperiment-class}
\alias{BiclusterExperiment-class}
\alias{as.matrix,BiclusterExperiment-method}
\alias{getStrat,BiclusterExperiment-method}
\alias{names,BiclusterExperiment-method}
\alias{wipe,BiclusterExperiment-method}
\alias{wipeExcept,BiclusterExperiment-method}
\alias{strategies,BiclusterExperiment-method}
\title{Class "BiclusterExperiment" for data and biclusters}
\usage{
\S4method{as.matrix}{BiclusterExperiment}(x)

\S4method{getStrat}{BiclusterExperiment}(bce, id)

\S4method{names}{BiclusterExperiment}(x)

\S4method{wipe}{BiclusterExperiment}(bce)

\S4method{wipeExcept}{BiclusterExperiment}(bce, bcs)

\S4method{strategies}{BiclusterExperiment}(bce)
}
\description{
This class encapsulates data for one or more biclustering runs derived from
the same abundance data. Objects can be created using the
\code{\link{BiclusterExperiment}} constructor. A subclass of
\code{\link[Biobase]{eSet-class}}.
}
\section{Methods (by generic)}{
\itemize{
\item \code{as.matrix}: Get abundance values in a BiclusterExperiment

\item \code{getStrat}: Get a BiclusterStrategy contained by a
BiclusterExperiment, by providing either name or integer index

\item \code{names}: Character names of BiclusterStrategies in this BiclusterExperiment

\item \code{wipe}: Return this BiclusterExperiment with all BiclusterStrategy objects removed

\item \code{wipeExcept}: Return this BiclusterExperiment with all
BiclusterStrategy objects removed except \code{bcs}. Argument \code{bcs}
can be passed as a name, integer index, or the BiclusterStrategy itself.

\item \code{strategies}: Get/set a list of the BiclusterStrategy objects
contained by this BiclusterExperiment.
}}

\section{Slots}{

\describe{
\item{\code{assayData}}{Object of class \code{\link[Biobase]{AssayData-class}} in the
"list" storage mode. \code{AssayData} must have a matrix named
"abund" with rows representing features and columns representing samples.
Any other matrices in assayData are ignored by this package.}

\item{\code{strategies}}{A \code{\link{list}} of
\code{\link{BiclusterStrategy-class}} objects}

\item{\code{phenoData}}{See \code{\link[Biobase]{eSet-class}}.}

\item{\code{featureData}}{\code{\link[Biobase]{eSet-class}}.}

\item{\code{experimentData}}{\code{\link[Biobase]{eSet-class}}. Not accessed by any 
functions in \code{\link{mfBiclust}}.}

\item{\code{annotation}}{\code{\link[Biobase]{eSet-class}}. Not accessed by any 
functions in \code{\link{mfBiclust}}.}

\item{\code{protocolData}}{\code{\link[Biobase]{eSet-class}}. Not accessed by any 
functions in \code{\link{mfBiclust}}.}
}}

\examples{
bce <- BiclusterExperiment(yeast_benchmark[[1]])
addStrat(bce, k = 3)
biclusterGUI(bce)
}
\seealso{
\code{\link[Biobase]{eSet}}
}
