% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiclusterExperiment.R, R/graphical.R
\docType{class}
\name{BiclusterExperiment-class}
\alias{BiclusterExperiment-class}
\alias{BiclusterExperiment,matrix-method}
\alias{BiclusterExperiment,ANY-method}
\alias{as.matrix,BiclusterExperiment-method}
\alias{clean,BiclusterExperiment-method}
\alias{getStrat,BiclusterExperiment-method}
\alias{names,BiclusterExperiment-method}
\alias{strategies,BiclusterExperiment-method}
\alias{wipe,BiclusterExperiment-method}
\alias{wipeExcept,BiclusterExperiment-method}
\alias{plot,BiclusterExperiment,ANY-method}
\title{Class "BiclusterExperiment" for data and biclusters}
\usage{
\S4method{BiclusterExperiment}{matrix}(m, bcs = list(),
  phenoData = Biobase::annotatedDataFrameFrom(m, byrow = FALSE),
  featureData = annotatedDataFrameFrom(m, byrow = TRUE), pp = FALSE,
  maxNa = 0)

\S4method{BiclusterExperiment}{ANY}(m, bcs = list(),
  phenoData = Biobase::annotatedDataFrameFrom(m, byrow = FALSE),
  featureData = annotatedDataFrameFrom(m, byrow = TRUE), pp = FALSE,
  maxNa = 0)

\S4method{as.matrix}{BiclusterExperiment}(x)

\S4method{clean}{BiclusterExperiment}(object, cleanParam = 0,
  dimsRemain = FALSE)

\S4method{getStrat}{BiclusterExperiment}(bce, id)

\S4method{names}{BiclusterExperiment}(x)

\S4method{strategies}{BiclusterExperiment}(bce)

\S4method{wipe}{BiclusterExperiment}(bce)

\S4method{wipeExcept}{BiclusterExperiment}(bce, bcs)

\S4method{plot}{BiclusterExperiment,ANY}(x, logBase = 0,
  phenoLabels = c(), biclustLabels = c(), ordering = c("input",
  "distance", "cluster"), strategy = "", rowNames = FALSE,
  colNames = FALSE)
}
\arguments{
\item{x}{a \code{\link{BiclusterExperiment-class}} object}

\item{object}{a BiclusterExperiment}

\item{cleanParam}{what proportion of data must be present}

\item{dimsRemain}{return extra information about remaining rows and columns}
}
\description{
This class encapsulates data for one or more biclustering runs derived from
the same abundance data. Objects can be created using the
\code{\link{BiclusterExperiment}} constructor. A subclass of
\code{\link{eSet}}.
}
\section{Methods (by generic)}{
\itemize{
\item \code{BiclusterExperiment}: default constuctor method

\item \code{BiclusterExperiment}: Attempts to coerce \code{m} to a matrix before
encapsulating it in a \code{BiclusterExperiment}

\item \code{as.matrix}: Get abundance values in a BiclusterExperiment

\item \code{clean}: Remove rows and columns with less than \code{cleanParam} entries present

\item \code{getStrat}: Get a BiclusterStrategy contained by a
BiclusterExperiment, by providing either name or integer index

\item \code{names}: Character names of BiclusterStrategies in this BiclusterExperiment

\item \code{strategies}: Get/set a list of the BiclusterStrategy objects
contained by this BiclusterExperiment.

\item \code{wipe}: Return this BiclusterExperiment with all BiclusterStrategy objects removed

\item \code{wipeExcept}: Return this BiclusterExperiment with all
BiclusterStrategy objects removed except \code{bcs}. Argument \code{bcs}
can be passed as a name, integer index, or the BiclusterStrategy itself.

\item \code{plot}: Create a heatmap from
\code{BiclusterExperiment@assayData}
}}

\section{Slots}{

\describe{
\item{\code{assayData}}{Object of class \code{\link{AssayData-class}} in the
"list" storage mode. \code{AssayData} must have a matrix named
"abund" with rows representing features and columns representing samples.
Any other matrices in assayData are ignored by this package.}

\item{\code{strategies}}{A \code{\link{list}} of
\code{\link{BiclusterStrategy-class}} objects}

\item{\code{phenoData}}{See \code{\link{eSet-class}}.}

\item{\code{featureData}}{\code{\link{eSet-class}}.}

\item{\code{experimentData}}{\code{\link{eSet-class}}. Not accessed by any
functions in \code{\link{mfBiclust}}.}

\item{\code{annotation}}{\code{\link{eSet-class}}. Not accessed by any
functions in \code{\link{mfBiclust}}.}

\item{\code{protocolData}}{\code{\link{eSet-class}}. Not accessed by any
functions in \code{\link{mfBiclust}}.}
}}

\examples{
bce <- BiclusterExperiment(yeast_benchmark[[1]])

}
\seealso{
\code{\link{eSet-class}}
}
