% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiclusterExperiment.R
\docType{class}
\name{BiclusterExperiment-class}
\alias{BiclusterExperiment-class}
\title{Class "BiclusterExperiment" for multiple biclustering results}
\description{
This class encapsulates factorization and thresholding data for one or more
biclustering runs. Objects can be created using the
\code{\link{BiclusterExperiment}} constructor.
}
\section{Slots}{

\describe{
\item{\code{data}}{Object of class \code{\link{matrix}}. The original data.}

\item{\code{annot}}{Object of class \code{\link{data.frame}}. Annotations provided by the user}

\item{\code{strategies}}{A \code{\link{list}} of \code{BiclusterStrategy} objects}
}}

