% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiclusterExperiment.R
\name{BiclusterExperiment}
\alias{BiclusterExperiment}
\title{Perform multiple biclustering runs}
\usage{
BiclusterExperiment(m, bcs = list(),
  phenoData = Biobase::annotatedDataFrameFrom(m, byrow = TRUE),
  featureData = annotatedDataFrameFrom(m, byrow = FALSE), pp = FALSE,
  maxNa = 0.5)
}
\arguments{
\item{m}{the data matrix defining this BiclusterExperiment. Should have rows
as samples and features as columns}
}
\value{
an instance of BiclusterExperiment-class
  containing the following slots, accessed with @: 
  data: Object of class \code{\link{matrix}}. The original data. 
  annot: Object of class \code{\link{data.frame}}. Annotations provided by the user
  strategies: A \code{\link{list}} of \code{BiclusterStrategy} objects
}
\description{
BiclusterExperiment constructs an object holding data from multiple
biclustering runs.
}
\details{
This function is useful for constructing one BiclusterExperiment
encapsulating results of different pipelines (e.g. comparing NMF with PCA).
For comparing results from the same pipeline with differing values of
\code{k}, \code{\link{mfbc()}} is recommended.
}
