% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiclusterExperiment.R
\name{BiclusterExperiment}
\alias{BiclusterExperiment}
\title{Construct a container for biclustering runs}
\usage{
BiclusterExperiment(m, bcs = list(),
  phenoData = Biobase::annotatedDataFrameFrom(m, byrow = FALSE),
  featureData = annotatedDataFrameFrom(m, byrow = TRUE), pp = FALSE,
  maxNa = 0.5)
}
\arguments{
\item{m}{the data matrix defining this BiclusterExperiment. Should have
samples as columns and features as rows.}
}
\description{
Constructs a \code{\link{BiclusterExperiment-class}} object holding data,
with a slot for future biclustering runs performed on that data. To run
biclustering, use \code{\link{addStrat}()}.
}
\examples{
#' bce <- BiclusterExperiment(yeast_benchmark[[1]])
#' addStrat(bce, k = 3)
#' biclusterGUI(bce)
}
