% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biclusterGUI.R
\docType{methods}
\name{biclusterGUI}
\alias{biclusterGUI}
\alias{biclusterGUI,BiclusterExperiment-method}
\alias{biclusterGUI,ExpressionSet-method}
\alias{biclusterGUI,missing-method}
\alias{biclusterGUI,ANY-method}
\title{Explore a BiclusterExperiment}
\usage{
biclusterGUI(obj = NULL, ...)

\S4method{biclusterGUI}{BiclusterExperiment}(obj = NULL, ...)

\S4method{biclusterGUI}{ExpressionSet}(obj = NULL, ...)

\S4method{biclusterGUI}{missing}(obj = NULL, ...)

\S4method{biclusterGUI}{ANY}(obj = NULL, ...)
}
\arguments{
\item{obj}{a BiclusterExperiment, an ExpressionSet, or an object
coercible to a matrix. If missing, the user must import data from a comma-
or space-delimited file using the GUI.}

\item{...}{Other parameters. \code{dbg = TRUE} Runs the GUI in a debug mode
that uses algorithm parameters tweaked to sacrifice accuracy for speed.}
}
\value{
A Shiny app object
}
\description{
Opens a shiny GUI to analyze and visualize bicluster analyses.
}
\section{Methods (by class)}{
\itemize{
\item \code{BiclusterExperiment}: Default method

\item \code{ExpressionSet}: Coerces \code{obj} to a
\code{\link{BiclusterExperiment-class}} object and runs GUI

\item \code{missing}: Runs GUI without pre-loading a dataset. Use the
"Data" tab to load data.

\item \code{ANY}: Attempts to encapsulate \code{obj} in a
\code{\link{BiclusterExperiment-class}} object and pass it to the GUI
}}

\examples{
\dontrun{biclusterGUI()
biclusterGUI(yeast_benchmark[[1]])
biclusterGUI(BiclusterExperiment(yeast_benchmark[[1]]))
}

}
