% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biclusterGUI.R
\name{biclusterGUI}
\alias{biclusterGUI}
\title{Explore a BiclusterExperiment}
\usage{
biclusterGUI(obj = NULL, ...)
}
\arguments{
\item{obj}{a BiclusterExperiment, an ExpressionSet, or an object
coercible to a matrix}

\item{dbg}{Runs in a debug mode that uses algorithm parameters increasing
sacrificing accuracy for speed.}
}
\description{
Opens a shiny GUI to visualize analysis results contained in a
BiclusterExperiment object
}
\examples{
bce <- BiclusterExperiment(yeast_benchmark[[1]])
addStrat(bce, k = 3)
biclusterGUI(bce)
}
