% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biclusterGUI.R
\docType{methods}
\name{biclusterGUI}
\alias{biclusterGUI}
\alias{biclusterGUI,ANY-method}
\title{Explore a BiclusterExperiment}
\usage{
biclusterGUI(obj = NULL, ...)

\S4method{biclusterGUI}{ANY}(obj, dbg = FALSE)
}
\description{
Opens a GUI to visualize analysis results contained in a BiclusterExperiment
object
}
\section{Methods (by class)}{
\itemize{
\item \code{ANY}: Open GUI for a BiclusterExperiment
}}

\examples{
#' bce <- BiclusterExperiment(yeast_benchmark[[1]])
#' addStrat(bce, k = 3)
#' biclusterGUI(bce)
}
