% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiclusterStrategy.R
\docType{class}
\name{BiclusterStrategy-class}
\alias{BiclusterStrategy-class}
\alias{bcNames,BiclusterStrategy-method}
\alias{featureFactor,BiclusterStrategy-method}
\alias{featureThresh,BiclusterStrategy-method}
\alias{sampleThresh,BiclusterStrategy-method}
\alias{method,BiclusterStrategy-method}
\alias{name,BiclusterStrategy-method}
\alias{nclust,BiclusterStrategy-method}
\alias{sampleFactor,BiclusterStrategy-method}
\alias{threshAlgo,BiclusterStrategy-method}
\title{Class "BiclusterStrategy" for biclusters}
\usage{
\S4method{bcNames}{BiclusterStrategy}(bcs, allBc = FALSE)

\S4method{featureFactor}{BiclusterStrategy}(bcs, allBc = FALSE)

\S4method{featureThresh}{BiclusterStrategy}(bcs, allBc = FALSE)

\S4method{sampleThresh}{BiclusterStrategy}(bcs, allBc = FALSE)

\S4method{method}{BiclusterStrategy}(bcs)

\S4method{name}{BiclusterStrategy}(bcs)

\S4method{nclust}{BiclusterStrategy}(bcs, allBc = FALSE)

\S4method{sampleFactor}{BiclusterStrategy}(bcs, allBc = FALSE)

\S4method{threshAlgo}{BiclusterStrategy}(bcs)
}
\arguments{
\item{bcs}{A BiclusterSrategy class}

\item{allBc}{Return data on all stored biclusters}
}
\description{
This class encapsulates bicluster results created by a call to
\code{\link{addStrat}()}.
}
\section{Methods (by generic)}{
\itemize{
\item \code{bcNames}: The names of biclusters in the
BiclusterStrategy

\item \code{featureFactor}: Retrieves \code{bcs@factors@fit@W}, the
matrix factor mapping features to biclusters

\item \code{featureThresh}: Retrieves \code{bcs@featureThresh}, the threshold(s)
used to determine each bicluster's feature membership

\item \code{sampleThresh}: Retrieves \code{bcs@sampleThresh}, the threshold(s)
used to determine each bicluster's sample membership

\item \code{method}: Retrieves \code{bcs@factors@method}, the
biclustering algorithm used to produce \code{bcs}

\item \code{name}: Gets a display-friendly name of a
BiclusterStrategy that includes its biclustering algorithm, its
thresholding algorithm, and the number of biclusters.

\item \code{nclust}: Number of clusters in a BiclusterStrategy

\item \code{sampleFactor}: Retrieves \code{bcs@factors@fit@H}, the
matrix factor mapping samples to biclusters

\item \code{threshAlgo}: The algorithm used to calculate biclustering
thresholds
}}

\section{Slots}{

\describe{
\item{\code{factors}}{a \code{\link[NMF]{NMFfit-class}} or
\code{\link{genericFit-class}} containing a pair of matrices \eqn{L_{m,k}}
and \eqn{R_{k,n}}.}

\item{\code{featureThresh}}{a vector of thresholds of length \eqn{k}}

\item{\code{sampleThresh}}{a vector of thresholds of length \eqn{k}}

\item{\code{threshAlgo}}{the name of the thresholding algorithm, or "user"}

\item{\code{biclust}}{a \code{\link{Biclust-class}} object containing
logical bicluster membership matrices and other information}

\item{\code{k}}{the number of biclusters to access when \code{allBc = FALSE}.}

\item{\code{name}}{a display friendly character string describing the object}
}}

\section{The \code{k} slot}{

The numbers of biclusters requested in the call to \code{addStrat()} is
stored in \code{BiclusterStrategy@k}. A \code{BiclusterStrategy-class}
object may store more than \code{k} biclusters, but accessors only return
data on the first \code{k} biclusters by default. To obtain data on all
stored biclusters, accessors must be called with \code{allBc = TRUE}.
}

\seealso{
\code{\link{getStrat}}
}
