% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiclusterStrategy.R
\docType{class}
\name{BiclusterStrategy-class}
\alias{BiclusterStrategy-class}
\alias{clusteredFeatures,BiclusterStrategy-method}
\alias{clusteredSamples,BiclusterStrategy-method}
\alias{nclust,BiclusterStrategy-method}
\alias{threshAlgo,BiclusterStrategy-method}
\title{Class "BiclusterStrategy" for biclusters}
\usage{
\S4method{clusteredFeatures}{BiclusterStrategy}(bcs, allBc = FALSE)

\S4method{clusteredSamples}{BiclusterStrategy}(bcs, allBc = FALSE)

\S4method{nclust}{BiclusterStrategy}(bcs, allBc = FALSE)

\S4method{threshAlgo}{BiclusterStrategy}(bcs)
}
\description{
This class encapsulates bicluster results for one biclustering algorithm, one
thresholding algorithm, and one bicluster quantity \eqn{k}. If the
biclustering algorithm returns more than biclusters than the value of
\code{k} provided to \code{\link{addStrat}()}, the extra biclusters will
be stored with a warning. See note on slot \code{k} below.
}
\section{Methods (by generic)}{
\itemize{
\item \code{clusteredFeatures}: A binary matrix showing feature-bicluster
membership

\item \code{clusteredSamples}: A binary matrix showing sample-bicluster
membership

\item \code{nclust}: Number of clusters in a BiclusterStrategy

\item \code{threshAlgo}: The algorithm used to calculate biclustering
thresholds
}}

\section{Slots}{

\describe{
\item{\code{factors}}{a \code{\link[NMF]{NMFfit-class}} or
\code{\link{genericFit-class}} containing a pair of matrices \eqn{L_{m,k}}
and \eqn{R_{k,n}}.}

\item{\code{scoreThresh}}{a vector of thresholds of length \eqn{k}}

\item{\code{loadingThresh}}{a vector of thresholds of length \eqn{k}}

\item{\code{threshAlgo}}{the name of the thresholding algorithm, or "user"}

\item{\code{biclust}}{a \code{\link[biclust]{Biclust-class}} object containing
logical bicluster membership matrices and other information}

\item{\code{k}}{the value of parameter \code{k} in \code{\link{addStrat}()}.
\code{k} causes accessor methods to hide any extra biclusters stored
in this object, unless said accessor methods are called with 
\code{allBc = TRUE}}

\item{\code{name}}{a display friendly character string describing the object}
}}

