% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{clusteredFeatures}
\alias{clusteredFeatures}
\title{Bicluster-feature matrix}
\usage{
clusteredFeatures(bcs, allBc = FALSE)
}
\description{
Gets the logical matrix where rows represent features and columns represent
biclusters.
}
