% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiclusterStrategy.R
\name{clusteredFeatures}
\alias{clusteredFeatures}
\title{Feature-bicluster clustering matrix}
\usage{
clusteredFeatures(bcs)
}
\description{
A binary matrix showing which features are members of which biclusters.
}
