% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{clusteredSamples}
\alias{clusteredSamples}
\title{Sample-bicluster matrix}
\usage{
clusteredSamples(bcs, allBc = FALSE)
}
\description{
Gets the logical matrix where rows represent biclusters and columns represent
samples.
}
