% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/BiclusterStrategy.R
\docType{methods}
\name{clusteredSamples}
\alias{clusteredSamples}
\alias{clusteredSamples,BiclusterStrategy-method}
\title{Sample-bicluster matrix}
\usage{
clusteredSamples(bcs, allBc = FALSE)

\S4method{clusteredSamples}{BiclusterStrategy}(bcs, allBc = FALSE)
}
\arguments{
\item{bcs}{A BiclusterStrategy}

\item{allBc}{TRUE to return data from all stored biclusters}
}
\value{
A logical matrix
}
\description{
Returns a logical matrix \eqn{B_{k x n}} where element \eqn{B_{i,j}} is TRUE
if sample \eqn{j} is in bicluster \eqn{i}. Elsewhere in \code{mfBiclust}
documentation, this format is called a "membership matrix".
}
\section{Methods (by class)}{
\itemize{
\item \code{BiclusterStrategy}: method for BiclusterStrategy objects
}}

\examples{
bce <- BiclusterExperiment(yeast_benchmark[[1]])
bce <- addStrat(bce, k = 2, method = "als-nmf")
bcs <- getStrat(bce, 1)
clusteredSamples(bcs)

}
