% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphical.R
\docType{methods}
\name{factorHeatmap}
\alias{factorHeatmap}
\alias{factorHeatmap,BiclusterExperiment,character-method}
\alias{factorHeatmap,BiclusterExperiment,numeric-method}
\alias{factorHeatmap,BiclusterExperiment,BiclusterStrategy-method}
\title{Plot a heatmap showing bicluster membership of samples or features}
\usage{
factorHeatmap(bce, bcs, type, ordering = "input", ...)

\S4method{factorHeatmap}{BiclusterExperiment,character}(bce, bcs, type,
  ordering = "input", ...)

\S4method{factorHeatmap}{BiclusterExperiment,numeric}(bce, bcs, type,
  ordering = "input", ...)

\S4method{factorHeatmap}{BiclusterExperiment,BiclusterStrategy}(bce, bcs,
  type = c("feature", "sample"), ordering = c("input", "distance",
  "cluster"), phenoLabels = c(), biclustLabels = c(),
  colNames = FALSE)
}
\arguments{
\item{bce}{A BiclusterExperiment object}

\item{bcs}{The name or index of a BiclusterStrategy contained by \code{bce},
or the BiclusterStrategy object itself}

\item{type}{either "feature" for feature-bicluster membership or "sample" for
sample-bicluster membership}

\item{ordering}{The default \code{ordering = "input"} preserves the order
of samples or features from \code{bce@assayData}. \code{"distance"}
reorders based on Euclidean distance calculated from \code{bce@assayData}.
\code{"cluster"} reorders based on bicluster membership.}

\item{...}{Optional arguments \code{phenoLabels}, \code{biclustLabels},
\code{ordering}, and \code{colNames}, described below:}

\item{phenoLabels}{an optional character vector of labels to annotate. If
\code{type = "feature"}, \code{phenoLabels} should be column names of
\code{Biobase::phenoData(bce)}}

\item{biclustLabels}{an optional character vector of labels to annotate.
Should be elements of \code{bcNames(bcs)}. Both \code{phenoLabels} and
\code{biclustLabels} may be specified.}

\item{colNames}{if \code{TRUE}, labels the samples/features}
}
\value{
a \code{\link[pheatmap]{pheatmap}-class} object
}
\description{
Reads data from \code{BiclusterStrategy@factors} to create a heatmap of
bicluster membership across all samples or features.
}
\section{Methods (by class)}{
\itemize{
\item \code{bce = BiclusterExperiment,bcs = character}: Plots a matrix factor from the
\code{\link{BiclusterStrategy-class}} object named \code{bcs} in
\code{bce@strategies}.

\item \code{bce = BiclusterExperiment,bcs = numeric}: Plots a matrix factor from the
\code{\link{BiclusterStrategy-class}} object at the index specified by
\code{bcs}

\item \code{bce = BiclusterExperiment,bcs = BiclusterStrategy}: Plots a matrix factor from \code{bcs}.
}}

\examples{
bce <- BiclusterExperiment(yeast_benchmark[[1]])
bce <- addStrat(bce, k = 2, method = "als-nmf")
bcs <- getStrat(bce, 1)
factorHeatmap(bce, bcs, type = "sample")

}
