% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{filter.biclust}
\alias{filter.biclust}
\title{Filter biclusters by overlap and quantity}
\usage{
filter.biclust(rowxBicluster, biclusterxCol, max = NULL, overlap = 0.25)
}
\arguments{
\item{rowxBicluster}{a binary matrix with samples as rows, biclusters as 
columns}

\item{biclusterxCol}{a binary matrix with samples as rows, biclusters as 
columns}

\item{max}{the maximum number of bicluster to keep}

\item{overlap}{the maximum portion of a bicluster that may overlap with
another bicluster}
}
\description{
Returns a pair of bicluster membership matrices and a list of the qualifying
biclusters.
}
