% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genBiclusters.R
\name{genSimData}
\alias{genSimData}
\title{Simulate a matrix with n biclusters}
\usage{
genSimData(n, overlapped = FALSE, file = "", striped = c("cols", ""),
  noise = 0.01, dynamSize = FALSE, dimx = 50, dimy = dimx)
}
\description{
Optional noise parameter
}
