% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{getStrat}
\alias{getStrat}
\title{Get a BiclusterStrategy}
\usage{
getStrat(bce, id)
}
\arguments{
\item{bce}{A \code{\linkS4class{BiclusterExperiment}} that contains one or
more \code{\linkS4class{BiclusterStrategy}} objects}

\item{id}{A character string or numeric index identifying the desired
\code{\linkS4class{BiclusterStrategy}}}
}
\value{
A \code{\link{BiclusterStrategy-class}} object
}
\description{
Returns one BiclusterStrategy in \code{bce} identified by \code{id}.
}
\examples{
bce <- BiclusterExperiment(yeast_benchmark[[1]])
bce <- addStrat(bce, k = 2)
getStrat(bce, "ALS-NMF | Otsu | 2")
getStrat(bce, 1)
}
