% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfBiclust.R
\docType{package}
\name{mfBiclust}
\alias{mfBiclust}
\alias{mfBiclust-package}
\title{Matrix-factorization biclustering toolbox}
\description{
Biclustering using PCA and non-negative matrix factorization
}
\details{
This package contains functions and classes for creating and analyzing
biclustering pipelines. Currently snmf with OTSU thresholding is supported.
Support for biclustering algorithms PCA, plaid, and bimax, and thresholding
algorithms ITA and FCM is planned.
}
\section{Functions}{

  \code{\link{biclusterGUI}}
  \code{\link{BiclusterExperiment}}
  \code{\link{BiclusterStrategy}}
}

\references{
\href{https://doi.org/10.1016/j.ymeth.2018.02.004}{Bi-clustering
  of metabolic data using matrix factorization tools}
}
\author{
Quan Gu, <quan.gu@glasgow.ac.uk>
  Maintainer: Jonathan Lim, <jlim7@comcast.net>
}
