% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiclusterStrategy.R
\name{name}
\alias{name}
\title{Name of a BiclusterStrategy}
\usage{
name(bcs)
}
\description{
Get this BiclusterStrategy's display-friendly name. If it does not have a name, computes a
string containing the biclustering algorithm, thresholding algorithm, and
number of biclusters in the given BiclusterStrategy.
}
\details{
This function may not be used to modify a BiclusterStrategy's name.
}
