% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/BiclusterStrategy.R
\docType{methods}
\name{name}
\alias{name}
\alias{name,list-method}
\title{Name of a BiclusterStrategy}
\usage{
name(bcs)

\S4method{name}{list}(bcs)
}
\arguments{
\item{bcs}{A \code{\link{BiclusterStrategy-class}} object}
}
\value{
character string
}
\description{
Gets a display-friendly name of a BiclusterStrategy that includes its
biclustering algorithm, its thresholding algorithm, and the number of
biclusters.
}
\details{
This function may not be used to modify a BiclusterStrategy's name.
}
\examples{
bce <- BiclusterExperiment(yeast_benchmark[[1]])
bce <- addStrat(bce, k = 2)
bcs <- getStrat(bce, 1)
name(bcs)
}
