% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphical.R
\docType{methods}
\name{plotDist}
\alias{plotDist}
\alias{plotDist,BiclusterExperiment-method}
\title{Distance heatmap}
\usage{
plotDist(bce, type, ordering = c("input", "distance", "cluster"),
  distType = c("euclidean", "pearson"), ...)

\S4method{plotDist}{BiclusterExperiment}(bce, type = c("sample",
  "feature"), ordering = c("input", "distance", "cluster"),
  distType = c("euclidean", "pearson"), phenoLabels = c(),
  biclustLabels = c(), strategy = "", rowColNames = FALSE)
}
\arguments{
\item{bce}{A \code{BiclusterExperiment} object}

\item{type}{"sample" or "feature"}

\item{ordering}{The default \code{ordering = "input"} preserves the order
of samples or features from \code{bce@assayData}. \code{"distance"}
reorders based on the distance calculated from \code{distType}.
\code{"cluster"} reorders based on bicluster membership.}

\item{distType}{The formula to use for distance}

\item{phenoLabels}{an optional character vector of labels to annotate. If
\code{type = "feature"}, \code{phenoLabels} should be column names of
\code{Biobase::phenoData(bce)}}

\item{biclustLabels}{an optional character vector of labels to annotate.
Should be elements of \code{bcNames(bcs)}. Both \code{phenoLabels} and
\code{biclustLabels} may be specified.}

\item{strategy}{The name or index of an object in \code{bce@strategies} from
which to obtain bicluster annotations, or a
\code{\link{BiclusterStrategy-class}} object.}

\item{rowColNames}{Should feature/sample names be printed?}
}
\value{
a \code{\link[pheatmap]{pheatmap}-class} object
}
\description{
Plot a heatmap of sample distance or feature distance calculated from the raw
data in a \code{\link{BiclusterExperiment-class}} object. Default is
Euclidean distance between samples.
}
\examples{
bce <- BiclusterExperiment(cancer_benchmark[[1]]$data)
plotDist(bce, type = "sample", ordering = "distance", rowColNames = TRUE)
plotDist(bce, type = "feature", distType = "pearson", ordering = "distance")

}
