% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphical.R
\docType{methods}
\name{plotDist,BiclusterExperiment-method}
\alias{plotDist,BiclusterExperiment-method}
\alias{plotDist}
\title{Distance heatmap}
\usage{
\S4method{plotDist}{BiclusterExperiment}(bce, type = c("samples", "features"),
  distType = c("euclidean", "pearson"), ordering = c("input", "distance",
  "cluster"), phenoLabels = c(), biclustLabels = c(), strategy = "",
  rowColNames = FALSE)
}
\description{
Plot a heatmap from distance values in the BiclusterExperiment object
}
