% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiclusterExperiment.R
\docType{methods}
\name{plotSamples}
\alias{plotSamples}
\alias{plotSamples,BiclusterExperiment-method}
\alias{plotSamples}
\title{Score plot}
\usage{
plotSamples(obj, thresholds = NULL, strategy, bicluster, ordering)

\S4method{plotSamples}{BiclusterExperiment}(obj, thresholds = NULL,
  strategy = "", bicluster = "Bicluster.1", ordering = c("input",
  "distance", "cluster"))
}
\arguments{
\item{thresholds}{A single numeric, vector, or matrix of thresholds. See 
documentation}
}
\description{
Plot cluster membership scores
}
