% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simdata}
\alias{simdata}
\title{Simulated bicluster datasets}
\format{An object of class \code{list} of length 4.}
\usage{
simdata
}
\description{
Simulated matrices containing biclusters and mild normally-distributed noise.
Generated using \code{genSimData}.
}
\keyword{datasets}
