% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bicluster.R
\name{svd_pca}
\alias{svd_pca}
\title{Wrapper for prcomp}
\usage{
svd_pca(m, k)
}
\arguments{
\item{m}{the target matrix}

\item{k}{the number of principal components}
}
\description{
Returns a \code{\link{genericFit}} object.
}
