% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bicluster.R
\name{svd_pca}
\alias{svd_pca}
\title{Principal component dimensionality reduction using SVD}
\usage{
svd_pca(A, k, ...)
}
\arguments{
\item{A}{the matrix to factorize}

\item{k}{the number of factors to compute}

\item{...}{Present for compatibility. \code{als_nmf} has no other
parameters}
}
\value{
a \code{\link{genericFit-class}} object
}
\description{
Factorizes matrix \code{A} as the product of score and loading matrices
respectively truncated to \code{k} rows and \code{k} columns.
}
