% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiclusterStrategy.R
\name{threshAlgo}
\alias{threshAlgo}
\title{Algorithm used to calculate threshold}
\usage{
threshAlgo(bcs)
}
\description{
Bicluster membership is determined by a binary thresholding function. The
value of the threshold can be either user-provided or calculated by the
Otsu algorithm.
}
