% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{threshold}
\alias{threshold}
\title{Apply threshold to a score or loading matrix}
\usage{
threshold(m, th, ...)
}
\description{
Returns a binary matrix of the same size as \code{m} where all elements over
the threshold are 1.
}
\details{
If th is a vector, the first element of th will be used as threshold for the
first col/row in m, etc.
}
