% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/BiclusterExperiment.R
\docType{methods}
\name{wipeExcept}
\alias{wipeExcept}
\alias{wipeExcept,BiclusterExperiment-method}
\title{Wipe all biclustering results except one}
\usage{
wipeExcept(bce, bcs)

\S4method{wipeExcept}{BiclusterExperiment}(bce, bcs)
}
\arguments{
\item{bce}{A \code{BiclusterExperiment-class} object}

\item{bcs}{A \code{BiclusterStrategy-class} object in \code{bce@strategies},
or the name or index of an object in \code{bce@strategies}}
}
\value{
A BiclusterExperiment object
}
\description{
Returns a BiclusterExperiment with the same data and metadata as \code{bce},
but with \code{BiclusterExperiment@strategies} containing the single
specified \code{BiclusterStrategy} instance. The \code{BiclusterStrategy} to
preserve can be specified by name, by index, or by reference to an identical
\code{BiclusterStrategy}. This function is useful for isolating your
"favorite" biclustering result.
}
\examples{
bce <- BiclusterExperiment(yeast_benchmark[[1]])
bce <- addStrat(bce, k = 2, method = "als-nmf")
bce <- addStrat(bce, k = 5, method = "als-nmf")
wipeExcept(bce, 2) # preserves the second BiclusterStrategy
}
