% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yeast_benchmark}
\alias{yeast_benchmark}
\title{Gene expression datasets for benchmarking detection of biological signal}
\format{A list of 17 data.frames where columns are samples and rows are
  genes, with ENSEMBL names}
\source{
\url{https://github.com/padilha/biclustlib/tree/master/}
}
\usage{
yeast_benchmark
}
\description{
Yeast microarray datasets originally published by Jaskowiak PA, Campello RJ,
Costa Filho IG (2013). All rownames are ENSEMBL gene names.
}
\details{
Redistributed from biclustlib: A Python library of biclustering algorithms
and evaluation measures. Copyright (C) 2017 Victor Alexandre Padilha.
}
\keyword{datasets}
